+++ src/layouts/default.jsx
@@ -21,6 +21,7 @@ const tabVisibility = {
 		"drive",
 		"quiz",
 		"users",
+		"handRaise",
 		"profile",
 		"settings",
 	],
@@ -30,6 +31,7 @@ const tabVisibility = {
 		"drive",
 		"quiz",
 		"users",
+		"handRaise",
 		"profile",
 		"settings",
 	],
@@ -41,6 +43,7 @@ const tabVisibility = {
 		"drive",
 		"quiz",
 		"users",
+		"handRaise",
 		"profile",
 		"settings",
 	],
@@ -342,7 +345,7 @@ const Default = ({ children }) => {
 							</>
 						)}
 					</Disclosure>
-					<div className="flex-1 h-navScreen overflow-x-hidden overflow-y-scroll scrollbar">
+					<div className="flex-1 h-navScreen overflow-x-hidden overflow-y-auto scrollbar">
 						{typeof children === "function"
 							? children()
 							: cloneElement(children, {})}