+++ src/pages/cohorts/[cohortId]/index.jsx
@@ -1,10 +1,13 @@
 import Session from "@/content/src/pages"
-import { useMeetingEmitter } from "@/plugins/src/sockets/emitters"
+import { useContentStore } from "@/content/src/store"
+import {
+	useHandRaiseEmitter,
+	useMeetingEmitter,
+} from "@/plugins/src/sockets/emitters"
 import { usePluginsSocketStore, usePluginsStore } from "@/plugins/src/store"
 import { classNames } from "@/web-core/src/utils"
 import {
 	Bars3CenterLeftIcon,
-	BarsArrowUpIcon,
 	ChevronRightIcon,
 	MagnifyingGlassIcon,
 	Squares2X2Icon,
@@ -14,6 +17,7 @@ import Link from "next/link"
 import { useRouter } from "next/router"
 import { useEffect, useMemo, useRef, useState } from "react"
 import { Back } from "~/components/atoms"
+import Loading from "~/components/atoms/Loading"
 import { Breadcrumb } from "~/components/molecules"
 import { Sessions } from "~/components/organisms"
 import {
@@ -112,8 +116,8 @@ const CurriculumList = ({ hierarchy }) => {
 	return (
 		<>
 			{isFetching ? (
-				<div className="flex h-full items-center justify-center">
-					Loading...
+				<div className="flex h-navScreen w-screen items-center justify-center">
+					<Loading />
 				</div>
 			) : (
 				<>
@@ -127,7 +131,7 @@ const CurriculumList = ({ hierarchy }) => {
 						</title>
 					</Head>
 					<div className="h-full flex-1 flex flex-col relative">
-						<div className="p-4 flex sm:flex-nowrap flex-wrap gap-4 justify-between items-center shadow-sm bg-neutral-50 dark:bg-neutral-900 sticky top-0 z-20 border-b-2 border-neutral-200 dark:border-neutral-700">
+						<div className="p-4 px-4 sm:px-6 lg:px-8 flex sm:flex-nowrap flex-wrap gap-4 justify-between items-center shadow-sm bg-neutral-50 dark:bg-neutral-900 sticky top-0 z-20 border-b-2 border-neutral-200 dark:border-neutral-700">
 							<div className="text-lg xl:max-w-1/2 w-full leading-6 font-medium flex space-x-2 items-center">
 								<Back />
 								<div
@@ -192,14 +196,15 @@ const CurriculumList = ({ hierarchy }) => {
 											name="search"
 											id="search"
 											autoComplete="off"
-											className="block md:min-w-40 w-full rounded-md border-0 bg-white/5 py-2 pl-10 pr-3.5 shadow-sm ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-purple-500 sm:text-sm sm:leading-6 rounded-r-none"
+											className="block md:min-w-40 w-full rounded-md border-0 bg-white/5 py-2 pl-10 pr-3.5 shadow-sm ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-purple-500 sm:text-sm sm:leading-6"
+											// rounded-r-none"
 											placeholder="Search"
 											onChange={e =>
 												setQuery(e.target.value)
 											}
 										/>
 									</div>
-									<button
+									{/* <button
 										type="button"
 										disabled
 										className="relative -ml-px inline-flex items-center gap-x-1.5 rounded-r-md px-3 py-2 text-sm font-semibold ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 bg-white/5 disabled:cursor-not-allowed"
@@ -209,7 +214,7 @@ const CurriculumList = ({ hierarchy }) => {
 											aria-hidden="true"
 										/>
 										Sort
-									</button>
+									</button> */}
 								</div>
 								<button
 									className="h-full p-2.5 ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 bg-white/5 rounded-md hover:shadow-md"
@@ -233,7 +238,7 @@ const CurriculumList = ({ hierarchy }) => {
 								</button>
 							</div>
 						</div>
-						<div className="@container/curriculum p-4 h-full overflow-y-scroll scrollbar">
+						<div className="@container/curriculum p-4 px-4 sm:px-6 lg:px-8 h-full overflow-y-scroll scrollbar">
 							<div
 								className={classNames(
 									"gap-4",
@@ -326,8 +331,8 @@ const SessionList = () => {
 	return (
 		<>
 			{isFetching ? (
-				<div className="flex h-full items-center justify-center">
-					Loading...
+				<div className="flex h-navScreen w-screen items-center justify-center">
+					<Loading />
 				</div>
 			) : (
 				<>
@@ -338,7 +343,7 @@ const SessionList = () => {
 						</title>
 					</Head>
 					<div className="h-full flex-1 flex flex-col relative">
-						<div className="p-4 flex sm:flex-nowrap flex-wrap gap-4 justify-between items-center shadow-sm bg-neutral-50 dark:bg-neutral-900 sticky top-0 z-20 border-b-2 border-neutral-200 dark:border-neutral-700">
+						<div className="p-4 px-4 sm:px-6 lg:px-8 flex sm:flex-nowrap flex-wrap gap-4 justify-between items-center shadow-sm bg-neutral-50 dark:bg-neutral-900 sticky top-0 z-20 border-b-2 border-neutral-200 dark:border-neutral-700">
 							<div className="text-lg xl:max-w-1/2 w-full leading-6 font-medium flex space-x-2 items-center">
 								<Back />
 								<div
@@ -391,14 +396,15 @@ const SessionList = () => {
 											name="search"
 											id="search"
 											autoComplete="off"
-											className="block md:min-w-40 w-full rounded-md border-0 bg-white/5 py-2 pl-10 pr-3.5 shadow-sm ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-purple-500 sm:text-sm sm:leading-6 rounded-r-none"
+											className="block md:min-w-40 w-full rounded-md border-0 bg-white/5 py-2 pl-10 pr-3.5 shadow-sm ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-purple-500 sm:text-sm sm:leading-6"
+											// rounded-r-none"
 											placeholder="Search"
 											onChange={e =>
 												setQuery(e.target.value)
 											}
 										/>
 									</div>
-									<button
+									{/* <button
 										type="button"
 										disabled
 										className="relative -ml-px inline-flex items-center gap-x-1.5 rounded-r-md px-3 py-2 text-sm font-semibold ring-1 ring-inset ring-neutral-300 dark:ring-neutral-700 bg-white/5 disabled:cursor-not-allowed"
@@ -408,11 +414,11 @@ const SessionList = () => {
 											aria-hidden="true"
 										/>
 										Sort
-									</button>
+									</button> */}
 								</div>
 							</div>
 						</div>
-						<div className="@container/session p-4 overflow-y-scroll scrollbar">
+						<div className="@container/session p-4 px-4 sm:px-6 lg:px-8 overflow-y-scroll scrollbar">
 							{filteredSessions && filteredSessions.length ? (
 								<div className="gap-4 grid @4xl/session:grid-cols-4 @md/session:grid-cols-3 @sm/session:grid-cols-2 grid-cols-1">
 									<Sessions sessions={filteredSessions} />
@@ -437,17 +443,25 @@ const CohortInfoWrapper = ({ children }) => {
 		disabled: currentHierarchy !== null,
 	})
 
-	if (isFetching) return <div>Loading...</div>
+	if (isFetching)
+		return (
+			<div className="w-full h-navScreen flex items-center justify-center">
+				<Loading />
+			</div>
+		)
 	return children
 }
 
 const CohortDataWrapper = () => {
 	const router = useRouter()
-	const { currentHierarchy, cohortTitle, orgName } = useStore(store => ({
-		currentHierarchy: store.currentHierarchy,
-		cohortTitle: store.cohortTitle,
-		orgName: store.orgInfo?.name ?? null,
-	}))
+	const { currentHierarchy, cohortTitle, orgName, dispatch } = useStore(
+		store => ({
+			currentHierarchy: store.currentHierarchy,
+			cohortTitle: store.cohortTitle,
+			orgName: store.orgInfo?.name ?? null,
+			dispatch: store.dispatch,
+		})
+	)
 
 	const hierarchy = useMemo(() => {
 		const hierarchyArr = hierarchyTypes[currentHierarchy]
@@ -459,6 +473,14 @@ const CohortDataWrapper = () => {
 		)
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [currentHierarchy])
+
+	useEffect(() => {
+		dispatch({
+			type: "SET_STATE",
+			payload: { pageHierarchy: hierarchy ?? "inSession" },
+		})
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [hierarchy])
 
 	return (
 		<>
@@ -473,6 +495,7 @@ const CohortDataWrapper = () => {
 					{ label: "Cohorts", to: "/cohorts" },
 					{ label: cohortTitle },
 				]}
+				className="border-t bg-neutral-50 dark:bg-neutral-900 z-50 sticky top-0"
 			/>
 			<div className="flex flex-1">
 				{hierarchy ? (
@@ -490,27 +513,73 @@ const CohortDataWrapper = () => {
 }
 
 const SessionView = () => {
-	const { user, currentHierarchy } = useStore(store => ({
+	const { user, orgInfo, currentHierarchy } = useStore(store => ({
 		user: store.user,
+		orgInfo: store.orgInfo,
 		currentHierarchy: store.currentHierarchy,
 	}))
-	const { connected, socket, roomId } = usePluginsSocketStore(store => ({
-		connected: store.meetingSocket.connected,
-		socket: store.meetingSocket.socket,
-		roomId: store.meetingSocket.roomId,
+	const {
+		meetingSocketConnected,
+		meetingSocket,
+		meetingSocketRoomId,
+		handRaiseSocketConnected,
+		handRaiseSocket,
+		handRaiseSocketRoomId,
+	} = usePluginsSocketStore(store => ({
+		meetingSocketConnected: store.meetingSocket.connected,
+		meetingSocket: store.meetingSocket.socket,
+		meetingSocketRoomId: store.meetingSocket.roomId,
+		handRaiseSocketConnected: store.handRaiseSocket.connected,
+		handRaiseSocket: store.handRaiseSocket.socket,
+		handRaiseSocketRoomId: store.handRaiseSocket.roomId,
 	}))
-	const { joinMeetingSocket } = useMeetingEmitter()
+	const hasEnded = useContentStore(store => store.session?.config?.hasEnded)
+
+	const { joinMeetingSocket, leave: leaveMeetingSocket } = useMeetingEmitter()
+	const { joinHandRaiseSocket, leave: leaveHandRaiseSocket } =
+		useHandRaiseEmitter()
 
 	useEffect(() => {
-		if (roomId === null && !connected && socket)
+		if (
+			meetingSocketRoomId === null &&
+			meetingSocketConnected &&
+			meetingSocket
+		)
 			joinMeetingSocket({ currentHierarchy })
+
+		return () => {
+			if (meetingSocketRoomId)
+				leaveMeetingSocket({ roomId: meetingSocketRoomId })
+		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [roomId, connected, socket])
+	}, [meetingSocketRoomId, meetingSocketConnected, meetingSocket])
+
+	useEffect(() => {
+		if (
+			handRaiseSocketRoomId === null &&
+			handRaiseSocketConnected &&
+			handRaiseSocket
+		)
+			joinHandRaiseSocket({ currentHierarchy })
+
+		return () => {
+			if (handRaiseSocketRoomId)
+				leaveHandRaiseSocket({ roomId: handRaiseSocketRoomId })
+		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [handRaiseSocketRoomId, handRaiseSocketConnected, handRaiseSocket])
 
 	return (
-		<div className="sm:pt-8 pt-4 flex-1 flex">
+		<div
+			className={classNames(
+				"relative sm:pt-8 pt-4 flex-1 flex",
+				hasEnded ? "sm:pb-8 pb-4" : ""
+			)}
+		>
+			<div className="gridLockBlueBeams -z-20 absolute inset-x-0 top-0 h-full bg-no-repeat dark:mix-blend-color bg-center" />
 			<Session
 				userData={user}
+				orgInfo={orgInfo}
 				currentHierarchy={currentHierarchy}
 			/>
 		</div>