+++ src/pages/cohorts/index.jsx
@@ -59,7 +59,10 @@ const Cohorts = ({ preview = false }) => {
 				<title>Cohorts - {orgName ?? "Lisa"}</title>
 			</Head>
 			{preview ? null : (
-				<Breadcrumb breadcrumbs={[{ label: "Cohorts" }]} />
+				<Breadcrumb
+					breadcrumbs={[{ label: "Cohorts" }]}
+					className="border-t bg-neutral-50 dark:bg-neutral-900 z-50 sticky top-0"
+				/>
 			)}
 			<div
 				className={classNames(
@@ -70,7 +73,7 @@ const Cohorts = ({ preview = false }) => {
 				<div
 					className={classNames(
 						"flex gap-4 justify-between items-center",
-						preview ? "p-4" : "sm:pt-8 pt-4"
+						preview ? "py-4" : "sm:pt-8 pt-4"
 					)}
 				>
 					<div className="flex gap-2">
@@ -92,7 +95,7 @@ const Cohorts = ({ preview = false }) => {
 					<div
 						className={classNames(
 							"@container/cohorts",
-							preview ? "p-4" : ""
+							preview ? "py-4" : ""
 						)}
 					>
 						<div className="gap-4 grid @2xl/cohorts:grid-cols-4 @md/cohorts:grid-cols-3 @sm/cohorts:grid-cols-2 grid-cols-1">
@@ -100,7 +103,8 @@ const Cohorts = ({ preview = false }) => {
 								?.slice(0, preview ? 5 : Infinity)
 								?.map(cohort => {
 									if (!cohort) return null
-									cohort = cohort.cohort
+									cohort = cohort?.cohort
+									if (!cohort?.type) return null
 									const cohortType = cohort.type
 										.map(t => t[0])
 										.join("")