+++ src/pages/index.jsx
@@ -4,6 +4,7 @@ import { Sessions } from "~/components/organisms"
 import { useGetDashboardSessions } from "~/hooks/queries/dashboard"
 import { useStore } from "~/store"
 import Cohorts from "./cohorts"
+import Loading from "~/components/atoms/Loading"
 
 const Home = () => {
 	const { user, dispatch } = useStore(store => ({
@@ -34,7 +35,10 @@ const Home = () => {
 
 	return (
 		<>
-			<Breadcrumb breadcrumbs={[]} />
+			<Breadcrumb
+				breadcrumbs={[]}
+				className="border-t bg-neutral-50 dark:bg-neutral-900 z-50 sticky top-0"
+			/>
 			<div className="@lg:px-8 px-2 sm:pt-8 pt-4 flex flex-col gap-8 justify-start">
 				<div className="flex flex-col gap-0.5">
 					<h1 className="text-2xl leading-8 font-semibold">
@@ -45,15 +49,17 @@ const Home = () => {
 					</p>
 				</div>
 				{fetchingUpcoming ? (
-					<>Loading...</>
+					<div className="w-full flex items-center justify-center">
+						<Loading className="w-10 h-10" />
+					</div>
 				) : upcoming.length ? (
 					<div className="flex flex-col rounded-md">
-						<p className="p-4 pb-0 text-sm leading-5 font-normal">
+						<p className="pt-4 text-sm leading-5 font-normal">
 							Upcoming Sessions
 						</p>
 
-						<div className="@container/session p-4">
-							<div className="gap-4 grid grid-flow-col @5xl:auto-cols-[calc(20%_-_32px)] @4xl:auto-cols-[calc(25%_-_32px)] @xl:auto-cols-[calc(33%_-_32px)] @sm:auto-cols-[calc(50%_-_32px)] auto-cols-[calc(75%_-_32px)] overflow-x-scroll scrollbar-hide">
+						<div className="@container/session py-4">
+							<div className="gap-4 grid grid-flow-col @5xl:auto-cols-[calc(25%_-_15px)] @4xl:auto-cols-[calc(25%_-_15px)] @xl:auto-cols-[calc(33%_-_32px)] @sm:auto-cols-[calc(52%_-_32px)] auto-cols-[calc(78%_-_32px)] overflow-x-scroll overflow-y-visible scrollbar-hide">
 								<Sessions
 									makeRoute
 									sessions={upcoming}
@@ -63,14 +69,16 @@ const Home = () => {
 					</div>
 				) : null}
 				{fetchingRecent ? (
-					<>Loading...</>
+					<div className="w-full flex items-center justify-center">
+						<Loading className="w-10 h-10" />
+					</div>
 				) : recent.length ? (
 					<div className="flex flex-col rounded-md">
-						<p className="p-4 pb-0 text-sm leading-5 font-normal">
+						<p className="pt-4 text-sm leading-5 font-normal">
 							Recent Sessions
 						</p>
-						<div className="@container/session p-4">
-							<div className="gap-4 grid grid-flow-col @5xl:auto-cols-[calc(20%_-_32px)] @4xl:auto-cols-[calc(25%_-_32px)] @xl:auto-cols-[calc(33%_-_32px)] @sm:auto-cols-[calc(50%_-_32px)] auto-cols-[calc(75%_-_32px)] overflow-x-scroll overflow-y-visible scrollbar-hide">
+						<div className="@container/session py-4">
+							<div className="gap-4 grid grid-flow-col @5xl:auto-cols-[calc(25%_-_15px)] @4xl:auto-cols-[calc(25%_-_15px)] @xl:auto-cols-[calc(33%_-_32px)] @sm:auto-cols-[calc(52%_-_32px)] auto-cols-[calc(78%_-_32px)] overflow-x-scroll overflow-y-visible scrollbar-hide">
 								<Sessions
 									makeRoute
 									sessions={recent}