+++ src/svgs/recorded.jsx
@@ -1,16 +1,25 @@
-import React from "react"
-
 const RecordedIcon = ({ className }) => {
 	return (
 		<>
 			<svg
 				xmlns="http://www.w3.org/2000/svg"
-				viewBox="0 0 512 512"
+				viewBox="0 0 102 80"
+				fill="none"
 				className={className}
 			>
 				<path
-					d="M448 32H361.9l-1 1-127 127h92.1l1-1L453.8 32.3c-1.9-.2-3.8-.3-5.8-.3zm64 128V96c0-15.1-5.3-29.1-14-40l-104 104H512zM294.1 32H201.9l-1 1L73.9 160h92.1l1-1 127-127zM64 32C28.7 32 0 60.7 0 96v64H6.1l1-1 127-127H64zM512 192H0V416c0 35.3 28.7 64 64 64H448c35.3 0 64-28.7 64-64V192z"
-					fill="currentColor"
+					d="M62.6957 47.4638L49.1075 38.405C46.2831 36.5221 42.5 38.5467 42.5 41.9412V60.0588C42.5 63.4533 46.2831 65.4779 49.1075 63.595L62.6957 54.5362C65.219 52.854 65.219 49.146 62.6957 47.4638Z"
+					stroke="currentColor"
+					strokeWidth="6"
+					strokeLinecap="round"
+					strokeLinejoin="round"
+				/>
+				<path
+					d="M89.25 51C89.25 72.1249 72.1249 89.25 51 89.25C29.8751 89.25 12.75 72.1249 12.75 51C12.75 29.8751 29.8751 12.75 51 12.75C72.1249 12.75 89.25 29.8751 89.25 51Z"
+					stroke="currentColor"
+					strokeWidth="6"
+					strokeLinecap="round"
+					strokeLinejoin="round"
 				/>
 			</svg>
 		</>